COMMERCE-5669 Move data set view to data.set.view.table package
