+++ modules/apps/commerce/commerce-pricing-web/src/main/java/com/liferay/commerce/pricing/web/internal/frontend/taglib/clay/data/set/view/table/CommerceDiscountChannelClayTableDataSetDisplayView.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.commerce.pricing.web.internal.frontend.taglib.clay.data.set.view;
+package com.liferay.commerce.pricing.web.internal.frontend.taglib.clay.data.set.view.table;
 
 import com.liferay.commerce.pricing.web.internal.frontend.constants.CommercePricingDataSetConstants;
 import com.liferay.frontend.taglib.clay.data.set.ClayDataSetDisplayView;
@@ -21,20 +21,19 @@ import com.liferay.frontend.taglib.clay.data.set.view.table.ClayTableSchema;
 import com.liferay.frontend.taglib.clay.data.set.view.table.ClayTableSchemaBuilder;
 import com.liferay.frontend.taglib.clay.data.set.view.table.ClayTableSchemaBuilderFactory;
 import com.liferay.frontend.taglib.clay.data.set.view.table.ClayTableSchemaField;
-import com.liferay.petra.string.StringPool;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Alessio Antonio Rendina
+ * @author Riccardo Alberti
  */
 @Component(
 	enabled = false, immediate = true,
-	property = "clay.data.set.display.name=" + CommercePricingDataSetConstants.COMMERCE_DATA_SET_KEY_PROMOTION_ENTRIES,
+	property = "clay.data.set.display.name=" + CommercePricingDataSetConstants.COMMERCE_DATA_SET_KEY_DISCOUNT_QUALIFIER_CHANNELS,
 	service = ClayDataSetDisplayView.class
 )
-public class CommercePromotionEntriesClayTableDataSetDisplayView
+public class CommerceDiscountChannelClayTableDataSetDisplayView
 	extends BaseTableClayDataSetDisplayView {
 
 	@Override
@@ -42,34 +41,11 @@ public class CommercePromotionEntriesClayTableDataSetDisplayView
 		ClayTableSchemaBuilder clayTableSchemaBuilder =
 			_clayTableSchemaBuilderFactory.create();
 
-		ClayTableSchemaField imageField =
+		ClayTableSchemaField nameClayTableSchemaField =
 			clayTableSchemaBuilder.addClayTableSchemaField(
-				"product.thumbnail", StringPool.BLANK);
+				"channel.name", "name");
 
-		imageField.setContentRenderer("image");
-
-		ClayTableSchemaField skuField =
-			clayTableSchemaBuilder.addClayTableSchemaField("sku.name", "sku");
-
-		skuField.setContentRenderer("actionLink");
-
-		clayTableSchemaBuilder.addClayTableSchemaField(
-			"product.name.LANG", "name");
-
-		clayTableSchemaBuilder.addClayTableSchemaField(
-			"sku.basePromoPriceFormatted", "base-promotion-price");
-
-		clayTableSchemaBuilder.addClayTableSchemaField(
-			"priceFormatted", "promotion-price");
-
-		clayTableSchemaBuilder.addClayTableSchemaField(
-			"discountLevelsFormatted", "unit-discount");
-
-		ClayTableSchemaField tieredPrice =
-			clayTableSchemaBuilder.addClayTableSchemaField(
-				"hasTierPrice", "tiered-price");
-
-		tieredPrice.setContentRenderer("boolean");
+		nameClayTableSchemaField.setContentRenderer("actionLink");
 
 		return clayTableSchemaBuilder.build();
 	}
