+++ modules/apps/commerce/commerce-pricing-web/src/main/java/com/liferay/commerce/pricing/web/internal/frontend/taglib/clay/data/set/view/table/CommercePriceListClayTableDataSetDisplayView.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.commerce.pricing.web.internal.frontend.taglib.clay.data.set.view;
+package com.liferay.commerce.pricing.web.internal.frontend.taglib.clay.data.set.view.table;
 
 import com.liferay.commerce.pricing.web.internal.frontend.constants.CommercePricingDataSetConstants;
 import com.liferay.frontend.taglib.clay.data.set.ClayDataSetDisplayView;
@@ -26,14 +26,14 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Riccardo Alberti
+ * @author Alessio Antonio Rendina
  */
 @Component(
 	enabled = false, immediate = true,
-	property = "clay.data.set.display.name=" + CommercePricingDataSetConstants.COMMERCE_DATA_SET_KEY_DISCOUNTS,
+	property = "clay.data.set.display.name=" + CommercePricingDataSetConstants.COMMERCE_DATA_SET_KEY_PRICE_LISTS,
 	service = ClayDataSetDisplayView.class
 )
-public class CommerceDiscountClayTableDataSetDisplayView
+public class CommercePriceListClayTableDataSetDisplayView
 	extends BaseTableClayDataSetDisplayView {
 
 	@Override
@@ -42,22 +42,30 @@ public class CommerceDiscountClayTableDataSetDisplayView
 			_clayTableSchemaBuilderFactory.create();
 
 		ClayTableSchemaField nameClayTableSchemaField =
-			clayTableSchemaBuilder.addClayTableSchemaField("title", "name");
+			clayTableSchemaBuilder.addClayTableSchemaField("name", "name");
 
 		nameClayTableSchemaField.setContentRenderer("actionLink");
+		nameClayTableSchemaField.setSortable(true);
 
-		clayTableSchemaBuilder.addClayTableSchemaField("target", "target");
-
-		clayTableSchemaBuilder.addClayTableSchemaField(
-			"amountFormatted", "amount");
+		clayTableSchemaBuilder.addClayTableSchemaField("author", "author");
 
 		clayTableSchemaBuilder.addClayTableSchemaField(
-			"level", "discount-level");
+			"catalogName", "catalog");
 
-		ClayTableSchemaField activeField =
-			clayTableSchemaBuilder.addClayTableSchemaField("active", "active");
+		clayTableSchemaBuilder.addClayTableSchemaField("priority", "priority");
 
-		activeField.setContentRenderer("boolean");
+		ClayTableSchemaField dateclayTableSchemaField =
+			clayTableSchemaBuilder.addClayTableSchemaField(
+				"createDate", "create-date");
+
+		dateclayTableSchemaField.setContentRenderer("date");
+		dateclayTableSchemaField.setSortable(true);
+
+		ClayTableSchemaField clayTableSchemaField =
+			clayTableSchemaBuilder.addClayTableSchemaField(
+				"workflowStatusInfo", "status");
+
+		clayTableSchemaField.setContentRenderer("status");
 
 		return clayTableSchemaBuilder.build();
 	}
