+++ modules/apps/commerce/commerce-pricing-web/src/main/java/com/liferay/commerce/pricing/web/internal/frontend/taglib/clay/data/set/view/table/CommerceProductPricingClassClayTableDataSetDisplayView.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.commerce.pricing.web.internal.frontend.taglib.clay.data.set.view;
+package com.liferay.commerce.pricing.web.internal.frontend.taglib.clay.data.set.view.table;
 
 import com.liferay.commerce.pricing.web.internal.frontend.constants.CommercePricingDataSetConstants;
 import com.liferay.frontend.taglib.clay.data.set.ClayDataSetDisplayView;
@@ -21,20 +21,19 @@ import com.liferay.frontend.taglib.clay.data.set.view.table.ClayTableSchema;
 import com.liferay.frontend.taglib.clay.data.set.view.table.ClayTableSchemaBuilder;
 import com.liferay.frontend.taglib.clay.data.set.view.table.ClayTableSchemaBuilderFactory;
 import com.liferay.frontend.taglib.clay.data.set.view.table.ClayTableSchemaField;
-import com.liferay.petra.string.StringPool;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Alessio Antonio Rendina
+ * @author Riccardo Alberti
  */
 @Component(
 	enabled = false, immediate = true,
-	property = "clay.data.set.display.name=" + CommercePricingDataSetConstants.COMMERCE_DATA_SET_KEY_PRICE_LIST_ENTRIES,
+	property = "clay.data.set.display.name=" + CommercePricingDataSetConstants.COMMERCE_DATA_SET_KEY_PRODUCT_PRICING_CLASSES,
 	service = ClayDataSetDisplayView.class
 )
-public class CommercePriceEntryClayTableDataSetDisplayView
+public class CommerceProductPricingClassClayTableDataSetDisplayView
 	extends BaseTableClayDataSetDisplayView {
 
 	@Override
@@ -42,34 +41,13 @@ public class CommercePriceEntryClayTableDataSetDisplayView
 		ClayTableSchemaBuilder clayTableSchemaBuilder =
 			_clayTableSchemaBuilderFactory.create();
 
-		ClayTableSchemaField imageField =
-			clayTableSchemaBuilder.addClayTableSchemaField(
-				"product.thumbnail", StringPool.BLANK);
+		ClayTableSchemaField nameClayTableSchemaField =
+			clayTableSchemaBuilder.addClayTableSchemaField("title", "name");
 
-		imageField.setContentRenderer("image");
-
-		ClayTableSchemaField skuField =
-			clayTableSchemaBuilder.addClayTableSchemaField("sku.name", "sku");
-
-		skuField.setContentRenderer("actionLink");
+		nameClayTableSchemaField.setContentRenderer("actionLink");
 
 		clayTableSchemaBuilder.addClayTableSchemaField(
-			"product.name.LANG", "name");
-
-		clayTableSchemaBuilder.addClayTableSchemaField(
-			"sku.basePriceFormatted", "base-price");
-
-		clayTableSchemaBuilder.addClayTableSchemaField(
-			"priceFormatted", "price-list-price");
-
-		clayTableSchemaBuilder.addClayTableSchemaField(
-			"discountLevelsFormatted", "unit-discount");
-
-		ClayTableSchemaField tieredPrice =
-			clayTableSchemaBuilder.addClayTableSchemaField(
-				"hasTierPrice", "tiered-price");
-
-		tieredPrice.setContentRenderer("boolean");
+			"numberOfProducts", "number-of-products");
 
 		return clayTableSchemaBuilder.build();
 	}
